+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- transform mandatory fields of IGS access platform and resources
+
 ### Changes
 
 ### Deprecated